feat: setting page (#18)
