+++ src/app/pages/index.ts
@@ -1,2 +1,3 @@
 export * from './profile-page';
 export * from './setting-page';
+export * from './user-input-page';
