+++ src/app/profile/page.tsx
@@ -1,0 +1,5 @@
+import { UserInputPage } from '@/app/pages';
+
+export default function Page() {
+  return <UserInputPage />;
+}
