+++ src/components/index.ts
@@ -1,1 +1,3 @@
 export * from './NavigationBar';
+export * from './VitalSection';
+export * from './OptionSection';
